refactor/move main components to mainComponent folder
