+++ src/components/mainComponents/BottomButtons/bottomButtons.scss
@@ -8,7 +8,7 @@
 }
 
 .previous-buttonMessages {
-  background-image: url('../../../public/VectorNextMessage.svg');
+  background-image: url('../../../../public/VectorNextMessage.svg');
   background-repeat: no-repeat;
   background-position: center;
   transform: scale(-1, -1);
@@ -17,7 +17,7 @@
 }
 
 .next-buttonMessages {
-  background-image: url('../../../public/VectorNextMessage.svg');
+  background-image: url('../../../../public/VectorNextMessage.svg');
   background-repeat: no-repeat;
   background-position: center;
   position: absolute;
