+++ src/components/mainComponents/Circle/circle.scss
