+++ src/components/mainComponents/DateCircle/DateCircle.jsx
@@ -1,6 +1,6 @@
 import './dateCircle.scss'
 import {useContext} from "react";
-import {PropsContext} from "../../context/context.jsx";
+import {PropsContext} from "../../../context/context.jsx";
 
 export default function DateCircle () {
     const {
