+++ src/components/mainComponents/Slider/slider.scss
@@ -38,13 +38,13 @@
   margin-bottom: 56px;
 }
 .previous-button {
-  background-image: url('../../../public/VectorPrevious.svg');
+  background-image: url('../../../../public/VectorPrevious.svg');
   background-repeat: no-repeat;
   background-position: center;
 }
 
 .next-button {
-  background-image: url('../../../public/VectorNext.svg');
+  background-image: url('../../../../public/VectorNext.svg');
   background-repeat: no-repeat;
   background-position: center;
 }
